Extract ExternalService typeclass that is private to ApiBasedExchange
ApiBasedExchange is suppoed to provide "exchange" functionality
combing calls to external clients and cache. It is not important which
client is being called. This was abstracted away by the indtroduction
of ExternalService.

Please note that ExternalService is typeclass that is not visible to
other elements in the system. It pops out at the "end of the world"
where modules are "glued" together.